feat: copy documents between spaces and local disk
