+++ packages/apps/composer-app/src/plugins/SpacePlugin/components/SpaceMain.tsx
@@ -4,17 +4,14 @@
 
 import React, { FC } from 'react';
 
-import { Document } from '@braneframe/types';
 import { useTextModel } from '@dxos/aurora-composer';
-import { isTypedObject, SpaceProxy } from '@dxos/client';
+import { SpaceProxy } from '@dxos/client';
 import { observer } from '@dxos/observable-object/react';
 import { useIdentity } from '@dxos/react-client';
 import { Surface, useGraphContext, useTreeView } from '@dxos/react-surface';
 
 import { SpacePlugin } from '../SpacePlugin';
-
-export const isDocument = (datum: unknown): datum is Document =>
-  isTypedObject(datum) && Document.type.name === datum.__typename;
+import { isDocument } from '../document';
 
 export const SpaceMain: FC<{}> = observer(() => {
   const identity = useIdentity();
